Introduced abstract functional test class EndToEnd
